Initial Fork of Sheet-Plus
Sheet-Plus is a module for Foundry that allows players to use their mobile devices
to see their basic inventory details. This is useful, but also has a lot of
items lacking, such as spells, stats, and the ability to modify their character sheet.

I would like to take a stab at adding some of this functionality, so I have forked the
branch as per the CopyLeft license attributed to it.