Set display of the board to none when canvas: hidden
I found-out that the previous solution was not working at all.
Because at the time the script to set the #board to display: none
was being run, the board has not been initiated at all.

Handled this via css instead of js so we can ensure no matter
how many times the board is created/modified/distroyed by other modules
the display will always be set to none